improve description of the item
directed the filter from 'description', which was not as
descriptive as it should, to 'content|encoded' where FayerWayer's RSS XML offers actual text from the post and thus, gives a more accurate description of the news. Changes will need to be made to the core for correct handling of this implementation.