Changing feed source
FayerWayer's feed was compliant and working but since it is an
AngularJS app itself, the only reasonable way of scraping the contents
of each post would be interacting with the same API that the ng app is
interacting, which, to be honest, was too much of hassle. Since I'm
a blogger myself I know Conectica.com (former Conectica.mx) is a great
tech news blog and is always up-to-date, plus parsing it is plain
simple with the DOMCrawler.

[ticket: X]